Update Ssd13xx project
